# server/openapi.yaml
openapi: 3.0.3
info:
  title: MyCalorie API
  version: "1.0.0"
  description: Backend API for the MyCalorie Flutter app.
servers:
  - url: https://mct.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local dev server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FoodSource:
      type: string
      description: >
        RECIPE marks the stand-in food behind a Recipe. Those carry the
        recipe's name for the diary and are never returned by /foods.
      enum: [OPEN_FOOD_FACTS, MYFCD, CUSTOM, RECIPE]
    Error:
      type: object
      properties:
        error:
          type: string
...
    FoodItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        caloriesPer100g:
          type: number
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        barcode:
          type: string
          nullable: true
        photoBase64:
          type: string
          nullable: true
          description: Raw base64-encoded image data (no data-URI prefix).
        source:
          $ref: "#/components/schemas/FoodSource"
        createdByUserId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        isRecent:
          type: boolean
          description: Only present on the no-search default list.
    FoodItemCreate:
      type: object
      required: [name, caloriesPer100g, proteinPer100g, carbsPer100g, fatPer100g]
      properties:
        name:
          type: string
          example: Nasi Lemak
        caloriesPer100g:
          type: number
          example: 200
        proteinPer100g:
          type: number
          example: 5
        carbsPer100g:
          type: number
          example: 25
        fatPer100g:
          type: number
          example: 9
        barcode:
          type: string
        photoBase64:
          type: string
    FoodImport:
      type: object
      description: >-
        Fixed shape for macro estimates worked out in a claude.ai chat (e.g.
        reading a label together, or calculating a meal from known
        components) and imported as a custom food without retyping. `source`
        here describes how the *chat* arrived at the numbers — unrelated to
        FoodItem's own `source` enum, which this endpoint always sets to
        CUSTOM.
      required: [name, estimatedTotalWeightG, macrosPer100g]
      properties:
        name:
          type: string
          example: Burrito Wrap (rice + chicken)
        estimatedTotalWeightG:
          type: number
          example: 300
        macrosPer100g:
          type: object
          required: [kcal, proteinG, carbsG, fatG]
          properties:
            kcal:
              type: number
            proteinG:
              type: number
            carbsG:
              type: number
            fatG:
              type: number
          example: { kcal: 175, proteinG: 14.7, carbsG: 17.3, fatG: 5.2 }
        totalMacros:
          type: object
          description: >-
            macrosPer100g × (estimatedTotalWeightG / 100). Optional, but if
            present must be within ~10% of that — a sanity check against
            obvious calculation errors, not treated as authoritative.
          properties:
            kcal:
              type: number
            proteinG:
              type: number
            carbsG:
              type: number
            fatG:
              type: number
          example: { kcal: 525, proteinG: 44, carbsG: 52, fatG: 15 }
        confidence:
          type: string
          example: "estimate - based on known component macros"
        source:
          type: string
          enum: [photo_estimate, label_read, manual_calculation]
        dateLogged:
          type: string
          format: date-time
          description: When known; informational only (not applied by this endpoint).
    FoodItemUpdate:
      type: object
      properties:
        name:
          type: string
        caloriesPer100g:
          type: number
        proteinPer100g:
          type: number
        carbsPer100g:
          type: number
        fatPer100g:
          type: number
        photoBase64:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /foods:
    get:
      summary: Search food items by name, or browse recents/available foods
      description: >-
        With `search`, does a case-insensitive substring match on name.
        Without it, returns the user's recently-logged foods first (each
        flagged `isRecent: true`), then fills the rest of the list up to 20
        with other available foods.
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching food items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"
    post:
      summary: Create a custom food item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /foods/import:
    post:
      summary: Import a food from a claude.ai macro-estimate JSON file
      description: >-
        Creates a CUSTOM food from the fixed macro-import shape (see
        FoodImport) — the file a user might export from a claude.ai chat
        after discussing a meal or nutrition label. Client sends the parsed
        file contents as the request body; there's no multipart upload.
        Returns the created FoodItem so the client can immediately offer to
        log it (user reviews and confirms, same as vision-log — nothing here
        creates a LogEntry).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodImport"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "400":
          description: Missing/invalid fields, or totalMacros doesn't match macrosPer100g × weight/100
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /foods/{id}:
    patch:
      summary: Edit a custom food's name, macros, and/or photo
      description: >-
        Only the food's creator may edit it. If any macro field changes, all
        existing LogEntry rows for this food are recalculated using the new
        per-100g values (fixing a data-entry mistake corrects history too).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "404":
          description: Not found, or not owned by this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a custom food
      description: >-
        Only the food's creator may delete it. Foods that already have log
        entries can't be deleted (the log history depends on them) — the
        request is rejected with 409 instead.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found, or not owned by this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Food has logged entries and can't be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /foods/barcode/{barcode}:
    get:
      summary: Look up a food item by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...