# server/src/routes/foods.js
const express = require("express");
const prisma = require("../db");
const { ensureQuickAddFood } = require("../quickAddFood");

const router = express.Router();

router.get("/", async (req, res) => {
  const search = req.query.search?.trim();

  // The quick-add placeholder food backs one-off calorie entries; never offer
  // it as something to log. Recipe stand-ins are excluded for the same reason,
  // and because their per-100g values are zero — logging one by serving size
  // would record nothing. Recipes are logged from the Recipes tab instead.
  const quickAddId = (await ensureQuickAddFood()).id;
  const notAPlaceholder = { id: { not: quickAddId }, source: { not: "RECIPE" } };

  if (search) {
    const foods = await prisma.foodItem.findMany({
      where: { name: { contains: search, mode: "insensitive" }, ...notAPlaceholder },
      orderBy: { name: "asc" },
      take: 50,
    });
    return res.json(foods);
  }

  // No search query: this is the screen's default view, so surface the
  // user's recently-logged foods first (for one-tap re-logging), then fill
  // the rest of the list up to 20 with other available foods.
  // `distinct` with the newest-first ordering means each row is that food's
  // most recent entry, so its serving and meal are the last ones used — which
  // the client needs to offer one-tap re-logging.
  const recentLogEntries = await prisma.logEntry.findMany({
    where: {
      userId: req.userId,
      foodItemId: { not: quickAddId },
      foodItem: { source: { not: "RECIPE" } },
    },
    orderBy: { loggedAt: "desc" },
    select: { foodItemId: true, servingGrams: true, mealType: true },
    distinct: ["foodItemId"],
    take: 20,
  });

  const recentFoodIds = recentLogEntries.map((entry) => entry.foodItemId);
  const lastUsedByFoodId = new Map(
    recentLogEntries.map((entry) => [
      entry.foodItemId,
      { lastServingGrams: entry.servingGrams, lastMealType: entry.mealType },
    ]),
  );
  const recentFoods = recentFoodIds.length
    ? await prisma.foodItem.findMany({ where: { id: { in: recentFoodIds } } })
    : [];

  // Prisma's `in` filter doesn't preserve order, so re-sort to match recency.
  const foodById = new Map(recentFoods.map((food) => [food.id, food]));
  const orderedRecentFoods = recentFoodIds.map((id) => foodById.get(id)).filter(Boolean);

  const remainingSlots = 20 - orderedRecentFoods.length;
  const otherFoods = remainingSlots > 0
    ? await prisma.foodItem.findMany({
        where: { id: { notIn: [...recentFoodIds, quickAddId] }, source: { not: "RECIPE" } },
        orderBy: { name: "asc" },
        take: remainingSlots,
      })
    : [];

  res.json([
    ...orderedRecentFoods.map((food) => ({
      ...food,
      isRecent: true,
      ...lastUsedByFoodId.get(food.id),
    })),
    ...otherFoods.map((food) => ({ ...food, isRecent: false })),
  ]);
});

router.get("/mine", async (req, res) => {
  const foods = await prisma.foodItem.findMany({
    where: { createdByUserId: req.userId },
    orderBy: { name: "asc" },
  });

  res.json(foods);
});

router.get("/barcode/:barcode", async (req, res) => {
  const food = await prisma.foodItem.findUnique({
    where: { barcode: req.params.barcode },
  });

  if (!food) {
    return res.status(404).json({ error: "No food found for this barcode" });
  }

  res.json(food);
});

router.post("/", async (req, res) => {
  const { name, caloriesPer100g, proteinPer100g, carbsPer100g, fatPer100g, barcode, photoBase64 } = req.body;

  if (!name || caloriesPer100g == null || proteinPer100g == null || carbsPer100g == null || fatPer100g == null) {
    return res.status(400).json({
      error: "name, caloriesPer100g, proteinPer100g, carbsPer100g, and fatPer100g are required",
    });
  }

  const food = await prisma.foodItem.create({
    data: {
      name,
      caloriesPer100g,
      proteinPer100g,
      carbsPer100g,
      fatPer100g,
      barcode: barcode || null,
      photoBase64: photoBase64 || null,
      source: "CUSTOM",
      createdByUserId: req.userId,
    },
  });

  res.status(201).json(food);
});

// Accepts the fixed macro-import JSON shape a user might generate while
// discussing a meal/label with Claude in a claude.ai chat (see docs), and
// creates it as a CUSTOM food — a lightweight alternative to typing the same
// numbers into the Custom Food form by hand. The client sends the already
//-parsed file contents as the request body (or any hand-built JSON matching
// the shape); there's no multipart upload involved.
const IMPORT_MACRO_FIELDS = ["kcal", "proteinG", "carbsG", "fatG"];

function validateFoodImport(body) {
  if (!body || typeof body !== "object" || Array.isArray(body)) {
    return "Request body must be a JSON object";
  }
  if (typeof body.name !== "string" || !body.name.trim()) {
    return "name is required and must be a non-empty string";
  }
  if (typeof body.estimatedTotalWeightG !== "number" || !(body.estimatedTotalWeightG > 0)) {
    return "estimatedTotalWeightG is required and must be a positive number";
  }
  if (!body.macrosPer100g || typeof body.macrosPer100g !== "object") {
    return "macrosPer100g is required";
  }
  for (const field of IMPORT_MACRO_FIELDS) {
    const value = body.macrosPer100g[field];
    if (typeof value !== "number" || Number.isNaN(value) || value < 0) {
      return `macrosPer100g.${field} is required and must be a non-negative number`;
    }
  }
  if (body.totalMacros !== undefined) {
    if (typeof body.totalMacros !== "object" || body.totalMacros === null || Array.isArray(body.totalMacros)) {
      return "totalMacros must be an object if present";
    }
    // Numbers here may carry rounding from chat-side math, so this is a
    // sanity check for obvious errors (e.g. a decimal-place slip), not an
    // exact-match requirement — tolerate ~10%, with a floor so tiny values
    // (near-zero fat, etc.) don't trip on rounding alone.
    const scale = body.estimatedTotalWeightG / 100;
    const mismatches = [];
    for (const field of IMPORT_MACRO_FIELDS) {
      const provided = body.totalMacros[field];
      if (typeof provided !== "number") continue;
      const expected = body.macrosPer100g[field] * scale;
      const tolerance = Math.max(expected * 0.1, field === "kcal" ? 5 : 1);
      if (Math.abs(provided - expected) > tolerance) {
        mismatches.push(`${field} (got ${provided}, expected ~${expected.toFixed(1)})`);
      }
    }
    if (mismatches.length > 0) {
      return `totalMacros doesn't match macrosPer100g × weight/100: ${mismatches.join(", ")}`;
    }
  }
  if (body.dateLogged !== undefined && Number.isNaN(new Date(body.dateLogged).getTime())) {
    return "dateLogged must be a valid date if present";
  }
  return null;
}

router.post("/import", async (req, res) => {
  const validationError = validateFoodImport(req.body);
  if (validationError) {
    return res.status(400).json({ error: validationError });
  }

  const { name, macrosPer100g } = req.body;

  const food = await prisma.foodItem.create({
    data: {
      name: name.trim(),
      caloriesPer100g: macrosPer100g.kcal,
      proteinPer100g: macrosPer100g.proteinG,
      carbsPer100g: macrosPer100g.carbsG,
      fatPer100g: macrosPer100g.fatG,
      source: "CUSTOM",
      createdByUserId: req.userId,
    },
  });

  res.status(201).json(food);
});

router.patch("/:id", async (req, res) => {
  const food = await prisma.foodItem.findUnique({ where: { id: req.params.id } });

  if (!food || food.createdByUserId !== req.userId) {
    return res.status(404).json({ error: "Custom food not found" });
  }

  const { name, caloriesPer100g, proteinPer100g, carbsPer100g, fatPer100g, photoBase64 } = req.body;
  const data = {};

  if (name !== undefined) data.name = name;
  if (caloriesPer100g !== undefined) data.caloriesPer100g = caloriesPer100g;
  if (proteinPer100g !== undefined) data.proteinPer100g = proteinPer100g;
  if (carbsPer100g !== undefined) data.carbsPer100g = carbsPer100g;
  if (fatPer100g !== undefined) data.fatPer100g = fatPer100g;
  if (photoBase64 !== undefined) data.photoBase64 = photoBase64;

  const macrosChanged = ["caloriesPer100g", "proteinPer100g", "carbsPer100g", "fatPer100g"].some(
    (key) => data[key] !== undefined,
  );

  const updated = await prisma.foodItem.update({ where: { id: food.id }, data });

  if (macrosChanged) {
    const entries = await prisma.logEntry.findMany({ where: { foodItemId: food.id } });
    await prisma.$transaction(
      entries.map((entry) => {
        const scale = entry.servingGrams / 100;
        return prisma.logEntry.update({
          where: { id: entry.id },
          data: {
            calories: updated.caloriesPer100g * scale,
            protein: updated.proteinPer100g * scale,
            carbs: updated.carbsPer100g * scale,
            fat: updated.fatPer100g * scale,
          },
        });
      }),
    );
  }

  res.json(updated);
});

router.delete("/:id", async (req, res) => {
  const food = await prisma.foodItem.findUnique({ where: { id: req.params.id } });

  if (!food || food.createdByUserId !== req.userId) {
    return res.status(404).json({ error: "Custom food not found" });
  }

  // LogEntry.foodItemId is a required relation, so deleting a food that has
  // been logged would violate the foreign key. Block it with a clear error
  // instead of letting Postgres throw.
  const logCount = await prisma.logEntry.count({ where: { foodItemId: food.id } });
  if (logCount > 0) {
    return res.status(409).json({ error: "This food has logged entries and can't be deleted" });
  }

  await prisma.foodItem.delete({ where: { id: food.id } });
  res.status(204).send();
});

module.exports = router;